add role form component and autodissmis alert
